+++ lua/neorg/modules/core/export/markdown/module.lua
@@ -14,7 +14,8 @@ like your markdown to be exported (i.e. do you want to support certain extension
 -- details will be necessary.
 
 local neorg = require("neorg.core")
-local module = neorg.modules.create("core.export.markdown")
+local modules = require("neorg.modules")
+local module = modules.create("core.export.markdown")
 
 module.setup = function()
     return {
