+++ lua/neorg/modules/core/export/module.lua
@@ -23,7 +23,8 @@ It takes 3 arguments:
 --]]
 
 local neorg = require("neorg.core")
-local module = neorg.modules.create("core.export")
+local modules = require("neorg.modules")
+local module = modules.create("core.export")
 
 module.setup = function()
     return {
@@ -35,7 +36,7 @@ module.setup = function()
 end
 
 module.load = function()
-    neorg.modules.await("core.neorgcmd", function(neorgcmd)
+    modules.await("core.neorgcmd", function(neorgcmd)
         neorgcmd.add_commands_from_table({
             export = {
                 args = 1,
@@ -69,14 +70,14 @@ module.public = {
     ---@param ftype string #The filetype to export to (as returned by e.g. `get_filetype()`)
     ---@return table,table #The export module and its configuration, else nil
     get_converter = function(ftype)
-        if not neorg.modules.is_module_loaded("core.export." .. ftype) then
-            if not neorg.modules.load_module("core.export." .. ftype) then
+        if not modules.is_module_loaded("core.export." .. ftype) then
+            if not modules.load_module("core.export." .. ftype) then
                 return
             end
         end
 
-        return neorg.modules.get_module("core.export." .. ftype),
-            neorg.modules.get_module_config("core.export." .. ftype)
+        return modules.get_module("core.export." .. ftype),
+            modules.get_module_config("core.export." .. ftype)
     end,
 
     --- Takes a buffer and exports it to a specific file
