+++ lua/neorg/modules/core/fs/module.lua
@@ -5,8 +5,8 @@
     ---
 --]]
 
-local neorg = require("neorg.core")
-local module = neorg.modules.create("core.fs")
+local modules = require("neorg.modules")
+local module = modules.create("core.fs")
 
 module.public = {
     directory_map = function(path, callback)
