+++ lua/neorg/modules/core/integrations/zen_mode/module.lua
@@ -1,7 +1,5 @@
-local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.integrations.zen_mode")
+local modules = require("neorg.modules")
+local module = modules.create("core.integrations.zen_mode")
 
 module.load = function()
     local success, zen_mode = pcall(require, "zen_mode")
