+++ lua/neorg/modules/core/mode/module.lua
@@ -14,10 +14,8 @@ which can be used by the user to switch modes.
 --]]
 
 local neorg = require("neorg.core")
-require("neorg.modules.base")
-require("neorg.events")
-
-local module = neorg.modules.create("core.mode")
+local modules = require("neorg.modules")
+local module = modules.create("core.mode")
 
 local log = neorg.log
 
@@ -61,8 +59,8 @@ module.public = {
         table.insert(module.private.modes, mode_name)
 
         -- Broadcast the mode_created event
-        neorg.events.broadcast_event(
-            neorg.events.create(
+        modules.events.broadcast_event(
+            modules.events.create(
                 module,
                 "core.mode.events.mode_created",
                 { current = module.config.public.current_mode, new = mode_name }
@@ -73,7 +71,7 @@ module.public = {
         table.insert(module.public.neorg_commands["mode"].complete[1], mode_name)
 
         -- If core.neorgcmd is loaded then update all autocompletions
-        local neorgcmd = neorg.modules.get_module("core.neorgcmd")
+        local neorgcmd = modules.get_module("core.neorgcmd")
 
         if neorgcmd then
             neorgcmd.sync()
@@ -99,8 +97,8 @@ module.public = {
         module.config.public.current_mode = mode_name
 
         -- Broadcast the mode_set event to all subscribed modules
-        neorg.events.broadcast_event(
-            neorg.events.create(
+        modules.events.broadcast_event(
+            modules.events.create(
                 module,
                 "core.mode.events.mode_set",
                 { current = module.config.public.previous_mode, new = mode_name }
@@ -143,8 +141,8 @@ module.on_event = function(event)
 end
 
 module.events.defined = {
-    mode_created = neorg.events.define(module, "mode_created"), -- Broadcast when a mode is created
-    mode_set = neorg.events.define(module, "mode_set"), -- Broadcast when a mode changes
+    mode_created = modules.events.define(module, "mode_created"), -- Broadcast when a mode is created
+    mode_set = modules.events.define(module, "mode_set"), -- Broadcast when a mode changes
 }
 
 module.events.subscribed = {
