+++ lua/neorg/modules/core/neorgcmd/commands/module/list/module.lua
@@ -7,10 +7,8 @@
 After module is loaded execute `:Neorg module list` to see a primitive list of currently loaded modules.
 --]]
 
-local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.neorgcmd.commands.module.list")
+local modules = require("neorg.modules")
+local module = modules.create("core.neorgcmd.commands.module.list")
 
 module.setup = function()
     return { success = true, requires = { "core.neorgcmd", "core.ui" } }
@@ -34,13 +32,13 @@ module.public = {
 module.on_event = function(event)
     if event.type == "core.neorgcmd.events.module.list" then
         local lines = {
-            -- neorg.modules.get_module_config("core.norg.concealer").icons.heading.level_1.icon
+            -- modules.get_module_config("core.norg.concealer").icons.heading.level_1.icon
             "*"
                 .. " "
                 .. "Loaded Neorg Modules",
         }
 
-        for _, mod in pairs(neorg.modules.loaded_modules) do
+        for _, mod in pairs(modules.loaded_modules) do
             table.insert(lines, "  - `" .. mod.name .. "`")
         end
