+++ lua/neorg/modules/core/norg/esupports/hop/module.lua
@@ -6,9 +6,8 @@
 --]]
 
 local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.norg.esupports.hop")
+local modules = require("neorg.modules")
+local module = modules.create("core.norg.esupports.hop")
 local job = require("plenary.job")
 
 module.setup = function()
