+++ lua/neorg/modules/core/norg/news/module.lua
@@ -6,7 +6,8 @@
 --]]
 
 local neorg = require("neorg.core")
-local module = neorg.modules.create("core.norg.news")
+local modules = require("neorg.modules")
+local module = modules.create("core.norg.news")
 
 local function is_version_greater_than(ver1, ver2)
     -- Here we assume that the versions aren't malformed
