+++ lua/neorg/modules/core/presenter/module.lua
@@ -6,9 +6,8 @@
 --]]
 
 local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.presenter")
+local modules = require("neorg.modules")
+local module = modules.create("core.presenter")
 local api = vim.api
 
 module.setup = function()
@@ -33,9 +32,9 @@ module.load = function()
     local keybinds = module.required["core.keybinds"]
 
     if module.config.public.zen_mode == "truezen" then
-        neorg.modules.load_module("core.integrations.truezen", module.name)
+        modules.load_module("core.integrations.truezen", module.name)
     elseif module.config.public.zen_mode == "zen-mode" then
-        neorg.modules.load_module("core.integrations.zen_mode", module.name)
+        modules.load_module("core.integrations.zen_mode", module.name)
     else
         neorg.log.error("Unrecognized mode for 'zen_mode' option. Please check your presenter config")
         error_loading = true
@@ -123,13 +122,13 @@ module.public = {
         end
 
         if
-            module.config.public.zen_mode == "truezen" and neorg.modules.is_module_loaded("core.integrations.truezen")
+            module.config.public.zen_mode == "truezen" and modules.is_module_loaded("core.integrations.truezen")
         then
-            neorg.modules.get_module("core.integrations.truezen").toggle_ataraxis()
+            modules.get_module("core.integrations.truezen").toggle_ataraxis()
         elseif
-            module.config.public.zen_mode == "zen-mode" and neorg.modules.is_module_loaded("core.integrations.zen_mode")
+            module.config.public.zen_mode == "zen-mode" and modules.is_module_loaded("core.integrations.zen_mode")
         then
-            neorg.modules.get_module("core.integrations.zen_mode").toggle()
+            modules.get_module("core.integrations.zen_mode").toggle()
         end
 
         -- Generate views selection popup
@@ -198,13 +197,13 @@ module.public = {
         module.required["core.mode"].set_mode(previous_mode)
 
         if
-            module.config.public.zen_mode == "truezen" and neorg.modules.is_module_loaded("core.integrations.truezen")
+            module.config.public.zen_mode == "truezen" and modules.is_module_loaded("core.integrations.truezen")
         then
-            neorg.modules.get_module("core.integrations.truezen").toggle_ataraxis()
+            modules.get_module("core.integrations.truezen").toggle_ataraxis()
         elseif
-            module.config.public.zen_mode == "zen-mode" and neorg.modules.is_module_loaded("core.integrations.zen_mode")
+            module.config.public.zen_mode == "zen-mode" and modules.is_module_loaded("core.integrations.zen_mode")
         then
-            neorg.modules.get_module("core.integrations.zen_mode").toggle()
+            modules.get_module("core.integrations.zen_mode").toggle()
         end
 
         api.nvim_buf_delete(module.private.buf, {})
