+++ lua/neorg/modules/core/storage/module.lua
@@ -6,10 +6,8 @@
     ---
 --]]
 
-local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.storage")
+local modules = require("neorg.modules")
+local module = modules.create("core.storage")
 
 module.setup = function()
     return {
