+++ lua/neorg/modules/core/ui/module.lua
@@ -7,9 +7,8 @@
 --]]
 
 local neorg = require("neorg.core")
-require("neorg.modules.base")
-
-local module = neorg.modules.create("core.ui")
+local modules = require("neorg.modules")
+local module = modules.create("core.ui")
 
 module.private = {
     windows = {},
