+++ lua/neorg/modules/core/ui/selection_popup.lua
@@ -1,8 +1,8 @@
 --[[
     A UI module to allow the user to press different keys to select different actions
 --]]
-local neorg = require("neorg.core")
-local module = neorg.modules.extend("core.ui.selection_popup")
+local modules = require("neorg.modules")
+local module = modules.extend("core.ui.selection_popup")
 
 module.private = {
     -- Stores all currently open selection popups
@@ -142,7 +142,7 @@ module.public = {
                 for _, key in ipairs(keys) do
                     -- TODO: Docs
                     local callback = function()
-                        neorg.modules
+                        modules
                             .get_module(module.name)
                             .invoke_key_in_selection(name, ({ key:gsub("<(.+)>", "|%1|") })[1], type)
                     end
@@ -178,7 +178,7 @@ module.public = {
                 for _, key in pairs(keys) do
                     -- TODO: Docs
                     local callback = function()
-                        neorg.modules
+                        modules
                             .get_module(module.name)
                             .invoke_key_in_selection(name, ({ key:gsub("<(.+)>", "|%1|") })[1], type)
                     end
